Kan hente ut query params for fnr, avtaletype og virksomhetsnummer
